Fix warning: implicit conversion changes signedness: 'unsigned int' to 'const int' [-Wsign-conversion]
In file included from visp/modules/core/test/image/testHSVGradient.cpp:45:
visp/modules/core/include/visp3/core/vpImageFilter.h:2605:26: warning: implicit conversion changes signedness: 'unsigned int' to 'const int' [-Wsign-conversion]
 2605 |     const int nbRows = I.getRows(), nbCols = I.getCols();
      |               ~~~~~~   ~~^~~~~~~~~
visp/modules/core/include/visp3/core/vpImageFilter.h:2087:7: note: in instantiation of function template specialization 'vpImageFilter::gradientFilterXMonothread<unsigned char, true, double>' requested here
 2087 |       gradientFilterXMonothread(I, GIx, filter, p_mask);
      |       ^
visp/modules/core/include/visp3/core/vpImageFilter.h:2429:5: note: in instantiation of function template specialization 'vpImageFilter::gradientFilterX<unsigned char, double, true>' requested here
 2429 |     gradientFilterX(I, GIx, nbThread, p_mask, type);
      |     ^

+++ modules/core/include/visp3/core/vpImageFilter.h
@@ -2602,8 +2602,8 @@ private:
     const vpImage<bool> *p_mask = nullptr
   )
   {
-    const int nbRows = I.getRows(), nbCols = I.getCols();
-    const int size = I.getSize();
+    const int nbRows = static_cast<int>(I.getRows()), nbCols = static_cast<int>(I.getCols());
+    const int size = static_cast<int>(I.getSize());
     const int offsetIdiff = nbCols;
 
     std::vector<OutputType> Idiff(size);
